Eslam♥

+++ API/Models/CustomModel/LastPrice.cs
@@ -8,7 +8,7 @@ namespace Inv.API.Models.CustomModel
     public class ModelLastPrice
     {
         public decimal CustLastPrice { get; set; }
-        public decimal CustLastTr { get; set; }
+        public int CustLastTr { get; set; }
         public string custLastDate { get; set; }
         public decimal LastPrice { get; set; }
         public decimal LastPurchase { get; set; }
